// src/types/generated/GlossaryDiff.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CategoryDiff } from "./CategoryDiff";

export type GlossaryDiff = { categories: Array<CategoryDiff>, total_added: number, total_removed: number, total_modified: number, has_changes: boolean, };